+++ clang-tools-extra/clang-tidy/bugprone/SignedCharMisuseCheck.cpp
@@ -16,6 +16,17 @@ using namespace clang::ast_matchers::internal;
 
 namespace clang::tidy::bugprone {
 
+namespace {
+
+AST_MATCHER(Stmt, isC23Stmt) {
+  return Finder->getASTContext().getLangOpts().C23;
+}
+AST_MATCHER(Decl, isC23Decl) {
+  return Finder->getASTContext().getLangOpts().C23;
+}
+
+} // anonymous namespace
+
 static constexpr int UnsignedASCIIUpperBound = 127;
 
 SignedCharMisuseCheck::SignedCharMisuseCheck(StringRef Name,
@@ -75,16 +86,21 @@ void SignedCharMisuseCheck::registerMatchers(MatchFinder *Finder) {
   const auto UnSignedCharCastExpr =
       charCastExpression(false, IntegerType, "unsignedCastExpression");
 
-  // Catch assignments with signed char -> integer conversion.
+  // Catch assignments with signed char -> integer conversion. Ignore false
+  // positives on C23 enums with the fixed underlying type of signed char.
   const auto AssignmentOperatorExpr =
       expr(binaryOperator(hasOperatorName("="), hasLHS(hasType(IntegerType)),
-                          hasRHS(SignedCharCastExpr)));
+                          hasRHS(SignedCharCastExpr)),
+           unless(allOf(isC23Stmt(), binaryOperator(hasLHS(hasType(
+                                         hasCanonicalType(enumType())))))));
 
   Finder->addMatcher(AssignmentOperatorExpr, this);
 
-  // Catch declarations with signed char -> integer conversion.
-  const auto Declaration = varDecl(isDefinition(), hasType(IntegerType),
-                                   hasInitializer(SignedCharCastExpr));
+  // Catch declarations with signed char -> integer conversion. Ignore false
+  // positives on C23 enums with the fixed underlying type of signed char.
+  const auto Declaration = varDecl(
+      isDefinition(), hasType(IntegerType), hasInitializer(SignedCharCastExpr),
+      unless(allOf(isC23Decl(), hasType(hasCanonicalType(enumType())))));
 
   Finder->addMatcher(Declaration, this);